@@ -1664,14 +1664,18 @@ class WebUser extends PimContext
     public function iPressTheButton($button, $modalWait = null)
     {
         $this->spin(function () use ($button, $modalWait) {
-            $this->getCurrentPage()->pressButton($button, true);
-
             if (null !== $modalWait) {
-                return null !== $this->getCurrentPage()->find('css', '.modal');
+                foreach ($this->getCurrentPage()->findAll('css', '.modal') as $modal) {
+                    if ($modal->isVisible()) {
+                        return true;
+                    }
+                }
             }
 
-            return true;
-        }, sprintf("Can not find any '%s' button", $button));
+            $this->getCurrentPage()->pressButton($button, true);
+
+            return null === $modalWait;
+        }, sprintf("Can not find any '%s' button%s", $button, null !== $modalWait ? ' or no modal found' : ''));
     }
 
     /**
@@ -2325,6 +2329,15 @@ class WebUser extends PimContext
     {
         $this->scrollContainerTo(900);
         $this->getCurrentPage()->select();
+    }
+
+    /**
+     * @Given /^I validate mass edit$/
+     */
+    public function iValidateTheMassEdit()
+    {
+        $this->scrollContainerTo(900);
+        $this->getCurrentPage()->confirm();
     }
 
     /**
