@@ -15,12 +15,11 @@ class SearchAfterSizeCursorFactorySpec extends ObjectBehavior
 {
     const DEFAULT_BATCH_SIZE = 100;
 
-    function let(Client $searchEngine, ObjectManager $om)
+    function let(Client $searchEngine, CursorableRepositoryInterface $cursorableRepository)
     {
         $this->beConstructedWith(
             $searchEngine,
-            $om,
-            ProductInterface::class,
+            $cursorableRepository,
             SearchAfterSizeCursor::class,
             self::DEFAULT_BATCH_SIZE,
             'pim_catalog_product'
@@ -33,9 +32,8 @@ class SearchAfterSizeCursorFactorySpec extends ObjectBehavior
         $this->shouldImplement('Akeneo\Component\StorageUtils\Cursor\CursorFactoryInterface');
     }
 
-    function it_creates_a_cursor($om, $searchEngine, CursorableRepositoryInterface $cursorableRepository)
+    function it_creates_a_cursor($searchEngine)
     {
-        $om->getRepository(ProductInterface::class)->willReturn($cursorableRepository);
         $searchEngine->search('pim_catalog_product', ['size' => 100, 'sort' => ['_uid' => 'asc'], 'search_after' => ['foo']])->willReturn([
             'hits' => [
                 'total' => 0,
@@ -47,9 +45,8 @@ class SearchAfterSizeCursorFactorySpec extends ObjectBehavior
             ->shouldBeAnInstanceOf('Akeneo\Component\StorageUtils\Cursor\CursorInterface');
     }
 
-    function it_creates_a_cursor_with_search_after_identifier($om, $searchEngine, CursorableRepositoryInterface $cursorableRepository)
+    function it_creates_a_cursor_with_search_after_identifier($searchEngine)
     {
-        $om->getRepository(ProductInterface::class)->willReturn($cursorableRepository);
         $searchEngine->search('pim_catalog_product', [
             'size' => 100, 'sort' => ['_uid' => 'asc'], 'search_after' => ['2017-12-12', 'pim_catalog_product#foo']
         ])->willReturn([
@@ -61,14 +58,5 @@ class SearchAfterSizeCursorFactorySpec extends ObjectBehavior
 
         $this->createCursor([], ['page_size' => 100, 'limit' => 150, 'search_after' => ['2017-12-12'], 'search_after_unique_key' => 'foo'])
             ->shouldBeAnInstanceOf('Akeneo\Component\StorageUtils\Cursor\CursorInterface');
-    }
-
-    function it_throws_an_exception_if_repository_is_not_cursorable($om)
-    {
-        $om->getRepository(ProductInterface::class)->willReturn(Argument::any());
-
-        $this->shouldThrow(
-            InvalidObjectException::objectExpected(ProductInterface::class, CursorableRepositoryInterface::class)
-        )->during('createCursor', [[], ['limit' => 150, 'search_after' => []]]);
     }
 }
