@@ -5,7 +5,6 @@ namespace Pim\Bundle\DataGridBundle\Datagrid\Configuration\Product;
 use Oro\Bundle\DataGridBundle\Datagrid\Common\DatagridConfiguration;
 use Oro\Bundle\DataGridBundle\Extension\Formatter\Configuration as FormatterConfiguration;
 use Pim\Bundle\DataGridBundle\Datagrid\Configuration\ConfiguratorInterface;
-use Pim\Component\Catalog\AttributeTypes;
 
 /**
  * Columns configurator for product grid, first column is identifier, then properties then ordered attributes
@@ -58,6 +57,7 @@ class ColumnsConfigurator implements ConfiguratorInterface
         $this->configuration = $configuration;
         $this->preparePropertiesColumns();
         $this->prepareAttributesColumns();
+        $this->prepareOtherColumns();
         $this->sortColumns();
         $this->addColumns();
     }
@@ -168,5 +168,20 @@ class ColumnsConfigurator implements ConfiguratorInterface
             sprintf(self::SOURCE_PATH, self::AVAILABLE_COLUMNS_KEY),
             $this->availableColumns
         );
+    }
+
+    private function prepareOtherColumns()
+    {
+        $otherColumns = $this->configuration->offsetGetByPath(
+            sprintf('[%s]', FormatterConfiguration::OTHER_COLUMNS_KEY)
+        );
+
+        if (null === $otherColumns) {
+            return;
+        }
+
+        foreach ($otherColumns as $columnCode => $columnData) {
+            $this->attributesColumns[$columnCode] = $columnData;
+        }
     }
 }
