@@ -6,10 +6,11 @@ use Akeneo\Component\StorageUtils\Cursor\CursorInterface;
 use Akeneo\Component\StorageUtils\Exception\InvalidObjectException;
 use Oro\Bundle\DataGridBundle\Datasource\ResultRecord;
 use Pim\Bundle\DataGridBundle\Extension\Pager\PagerExtension;
+use Pim\Component\Catalog\Model\AssociationInterface;
 use Pim\Component\Catalog\Model\ProductInterface;
+use Pim\Component\Catalog\Model\ProductModelInterface;
 use Pim\Component\Catalog\Query\Filter\Operators;
 use Pim\Component\Catalog\Query\ProductQueryBuilderInterface;
-use Pim\Component\Catalog\Query\Sorter\Directions;
 
 /**
  * Product datasource dedicated to the product association datagrid.
@@ -50,54 +51,77 @@ class AssociatedProductDatasource extends ProductDatasource
      */
     public function getResults()
     {
-        $currentProduct = $this->getConfiguration('current_product', false);
-        if (!$currentProduct instanceof ProductInterface) {
-            throw InvalidObjectException::objectExpected($currentProduct, ProductInterface::class);
+        $sourceProduct = $this->getConfiguration('current_product', false);
+        if (!$sourceProduct instanceof ProductInterface) {
+            throw InvalidObjectException::objectExpected($sourceProduct, ProductInterface::class);
         }
 
-        $associatedProductsIdentifiers = $this->getAssociatedProductsIdentifiers(
-            $currentProduct,
-            $this->getConfiguration('association_type_id')
-        );
+        $association = $this->getAssociation($sourceProduct, $this->getConfiguration('association_type_id'));
+        if (null === $association) {
+            return ['totalRecords' => 0, 'data' => []];
+        }
+
+        $associatedProductsIdentifiers = $this->getAssociatedProductIdentifiers($association);
+        $associatedProductModelsIdentifiers = $this->getAssociatedProductModelIdentifiers($association);
 
         $limit = (int)$this->getConfiguration(PagerExtension::PER_PAGE_PARAM, false);
         $locale = $this->getConfiguration('locale_code');
         $scope = $this->getConfiguration('scope_code');
         $from = null !== $this->getConfiguration('from', false) ?
-            (int)$this->getConfiguration('from', false) : 0;
-
-        $productCursor = $this->pqb->execute();
+            (int) $this->getConfiguration('from', false) : 0;
 
         $associatedProducts = $this->getAssociatedProducts(
             $associatedProductsIdentifiers,
             $limit,
-            0,
+            $from,
             $locale,
             $scope
         );
 
-        $rows = ['totalRecords' => count($associatedProducts)];
-        $rows['data'] = $associatedProducts;
+        $productModelLimit = $limit - count($associatedProducts);
+        $associatedProductModels = [];
+        if ($productModelLimit > 0) {
+            $productModelFrom = $from - count($associatedProductsIdentifiers) + count($associatedProducts);
+            $associatedProductModels = $this->getAssociatedProductModels(
+                $associatedProductModelsIdentifiers,
+                $productModelLimit,
+                max($productModelFrom, 0),
+                $locale,
+                $scope
+            );
+        }
+
+        $rows = ['totalRecords' => count($associatedProductsIdentifiers) + count($associatedProductModelsIdentifiers)];
+        $rows['data'] = array_merge($associatedProducts, $associatedProductModels);
 
         return $rows;
     }
 
     /**
-     * @param ProductInterface $product
-     * @param string           $associationTypeId
+     * @param AssociationInterface $association
      *
      * @return string[]
      */
-    protected function getAssociatedProductsIdentifiers(ProductInterface $product, $associationTypeId)
+    protected function getAssociatedProductIdentifiers(AssociationInterface $association): array
     {
         $identifiers = [];
+        foreach ($association->getProducts() as $associatedProduct) {
+            $identifiers[] = $associatedProduct->getIdentifier();
+        }
 
-        foreach ($product->getAssociations() as $association) {
-            if ($association->getAssociationType()->getId() === (int)$associationTypeId) {
-                foreach ($association->getProducts() as $associatedProduct) {
-                    $identifiers[] = $associatedProduct->getIdentifier();
-                }
-            }
+        return $identifiers;
+    }
+
+    /**
+     * @param AssociationInterface $association
+     *
+     * @return string[]
+     */
+    protected function getAssociatedProductModelIdentifiers(AssociationInterface $association): array
+    {
+        $identifiers = [];
+        foreach ($association->getProductModels() as $associatedProduct) {
+            $identifiers[] = $associatedProduct->getCode();
         }
 
         return $identifiers;
@@ -121,47 +145,69 @@ class AssociatedProductDatasource extends ProductDatasource
     ) {
         $pqb = $this->createQueryBuilder($limit, $from, $locale, $scope);
         $pqb->addFilter('identifier', Operators::IN_LIST, $associatedProductsIdentifiers);
-
+        $pqb->addFilter('entity_type', Operators::EQUALS, ProductInterface::class);
         $products = $pqb->execute();
 
-        return $this->normalizeProducts($products, $associatedProductsIdentifiers, $locale, $scope);
+        return $this->normalizeProductsAndProductModels($products, $locale, $scope);
+    }
+
+    /**
+     * @param array  $associatedProductModelsIdentifiers
+     * @param int    $limit
+     * @param int    $from
+     * @param string $locale
+     * @param string $scope
+     *
+     * @return array
+     */
+    protected function getAssociatedProductModels(
+        array $associatedProductModelsIdentifiers,
+        $limit,
+        $from,
+        $locale,
+        $scope
+    ) {
+        $pqb = $this->createQueryBuilder($limit, $from, $locale, $scope);
+        $pqb->addFilter('identifier', Operators::IN_LIST, $associatedProductModelsIdentifiers);
+        $pqb->addFilter('entity_type', Operators::EQUALS, ProductModelInterface::class);
+        $products = $pqb->execute();
+
+        return $this->normalizeProductsAndProductModels($products, $locale, $scope);
     }
 
     /**
      * @param CursorInterface $products
-     * @param string[]        $associatedProductsIdentifiers
      * @param string          $locale
      * @param string          $scope
      *
      * @return array
      */
-    protected function normalizeProducts(
+    protected function normalizeProductsAndProductModels(
         CursorInterface $products,
-        array $associatedProductsIdentifiers,
         $locale,
         $scope
     ) {
         $dataLocale = $this->getParameters()['dataLocale'];
 
         $context = [
-            'locales'     => [$locale],
-            'channels'    => [$scope],
-            'data_locale' => $dataLocale,
+            'locales'       => [$locale],
+            'channels'      => [$scope],
+            'data_locale'   => $dataLocale,
+            'is_associated' => true,
         ];
 
         $data = [];
         foreach ($products as $product) {
-            $context['is_associated'] = in_array($product->getIdentifier(), $associatedProductsIdentifiers);
-
-            $normalizedProduct = array_merge(
+            $normalized = array_merge(
                 $this->normalizer->normalize($product, 'datagrid', $context),
                 [
                     'id'         => $product->getId(),
                     'dataLocale' => $dataLocale,
+                    'is_associated' => true,
                 ]
             );
 
-            $data[] = new ResultRecord($normalizedProduct);
+            $data[] = new ResultRecord($normalized);
         }
 
         return $data;
@@ -202,5 +248,21 @@ class AssociatedProductDatasource extends ProductDatasource
         $pqb = $this->factory->create($factoryConfig);
 
         return $pqb;
+    }
+
+    /**
+     * @param ProductInterface           $sourceProduct
+     * @param mixed                      $associationTypeId
+     * @return null|AssociationInterface
+     */
+    private function getAssociation(ProductInterface $sourceProduct, $associationTypeId): ?AssociationInterface
+    {
+        foreach ($sourceProduct->getAssociations() as $association) {
+            if ($association->getAssociationType()->getId() === (int)$associationTypeId) {
+                return $association;
+            }
+        }
+
+        return null;
     }
 }
