@@ -41,6 +41,7 @@ class ProductAssociationNormalizer implements NormalizerInterface, SerializerAwa
 
         $data = [];
         $locale = current($context['locales']);
+        $channel = current($context['channels']);
 
         $data['identifier'] = $product->getIdentifier();
         $data['family'] = $this->getFamilyLabel($product, $locale);
@@ -50,7 +51,7 @@ class ProductAssociationNormalizer implements NormalizerInterface, SerializerAwa
 
         $data['is_checked'] = $context['is_associated'];
         $data['is_associated'] = $context['is_associated'];
-        $data['label'] = $product->getLabel($locale);
+        $data['label'] = $product->getLabel($locale, $channel);
         $data['completeness'] = $this->getCompleteness($product, $context);
         $data['image'] = $this->imageNormalizer->normalize($product->getImage(), $context['data_locale']);
 
