@@ -11,6 +11,7 @@ define(
     [
         'underscore',
         'oro/translator',
+        'pim/router',
         'pim/form',
         'pim/user-context',
         'pim/template/product/form/product-completeness'
@@ -18,12 +19,13 @@ define(
     function (
         _,
         __,
+        router,
         BaseForm,
         UserContext,
         template
     ) {
         return BaseForm.extend({
-            className: 'AknDropdown',
+            className: 'AknDropdown AknButtonList-item',
             template: _.template(template),
             events: {
                 'click .missing-attribute': 'showAttribute'
@@ -151,6 +153,27 @@ define(
                         context: 'base_product'
                     }
                 );
+
+                const product = this.getFormData();
+                const familyVariant = product.meta.family_variant;
+                const attributeCode = event.currentTarget.dataset.attribute;
+
+                if (null !== familyVariant) {
+                    if (!product.meta.attributes_for_this_level.includes(attributeCode)) {
+                        let modelId = product.meta.variant_navigation[0].selected.id;
+                        const comesFromParent = product.meta.parent_attributes.includes(attributeCode);
+                        const hasTwoLevelsOfVariation = (3 === product.meta.variant_navigation.length);
+
+                        if (comesFromParent && hasTwoLevelsOfVariation) {
+                            modelId = product.meta.variant_navigation[1].selected.id;
+                        }
+
+                        this.redirectToModel(modelId);
+
+                        return;
+                    }
+                }
+
                 this.getRoot().trigger(
                     'pim_enrich:form:show_attribute',
                     {
@@ -160,6 +183,20 @@ define(
                     }
                 );
                 this.render();
+            },
+
+            /**
+             * Redirect to the product model with the modelId
+             *
+             * @param modelId
+             */
+            redirectToModel: function(modelId) {
+                const params = {id: modelId};
+                const route = 'pim_enrich_product_model_edit';
+
+                sessionStorage.setItem('filter_missing_required_attributes', true);
+
+                router.redirectToRoute(route, params);
             }
         });
     }
