@@ -79,13 +79,14 @@ class ProductAndProductModelQueryBuilderSpec extends ObjectBehavior
             ]
         );
 
+        $pqb->addFilter('parent', Operators::IS_EMPTY, null, [])->shouldNotBeCalled();
         $pqb->addFilter('attributes_for_this_level', Operators::IN_LIST, ['foo', 'baz'], [])->shouldBeCalled();
         $pqb->execute()->willReturn($cursor);
 
         $this->execute()->shouldReturn($cursor);
     }
 
-    function it_executes_the_query_by_adding_a_default_filter_on_parents_when_there_is_no_attribute_filter_nor_parent_filter($pqb, CursorInterface $cursor)
+    function it_executes_the_query_by_adding_a_default_filter_on_parents($pqb, CursorInterface $cursor)
     {
         $pqb->getRawFilters()->willReturn(
             [
@@ -106,31 +107,31 @@ class ProductAndProductModelQueryBuilderSpec extends ObjectBehavior
         $this->execute()->shouldReturn($cursor);
     }
 
-    function it_executes_the_query_by_adding_a_filter_on_attributes_and_filter_on_parent($pqb, CursorInterface $cursor)
+    function it_does_not_add_a_default_filter_on_parents_when_there_is_a_source_attribute_filter($pqb, CursorInterface $cursor)
     {
         $pqb->getRawFilters()->willReturn(
             [
-                [
-                    'field'    => 'foo',
-                    'operator' => 'CONTAINS',
-                    'value'    => '42',
-                    'context'  => [],
-                    'type'     => 'attribute'
-                ],
                 [
                     'field'    => 'bar',
                     'operator' => 'IN LIST',
                     'value'    => ['toto'],
                     'context'  => [],
-                    'type'     => 'field'
-                ],
-                [
-                    'field'    => 'baz',
-                    'operator' => 'EQUALS',
-                    'value'    => 'sku_893042',
-                    'context'  => [],
                     'type'     => 'attribute'
                 ],
+            ]
+        );
+
+        $pqb->addFilter('parent', Argument::cetera())->shouldNotBeCalled();
+        $pqb->addFilter('attributes_for_this_level', Argument::cetera())->shouldBeCalled();
+        $pqb->execute()->willReturn($cursor);
+
+        $this->execute()->shouldReturn($cursor);
+    }
+
+    function it_does_not_add_a_default_filter_on_parents_when_there_is_a_source_parent_filter($pqb, CursorInterface $cursor)
+    {
+        $pqb->getRawFilters()->willReturn(
+            [
                 [
                     'field'    => 'parent',
                     'operator' => 'IN LIST',
@@ -141,8 +142,107 @@ class ProductAndProductModelQueryBuilderSpec extends ObjectBehavior
             ]
         );
 
-        $pqb->addFilter('parent', Operators::IS_EMPTY, null, [])->shouldNotBeCalled();
-        $pqb->addFilter('attributes_for_this_level', Operators::IN_LIST, ['foo', 'baz'], [])->shouldBeCalled();
+        $pqb->addFilter('parent', Argument::cetera())->shouldNotBeCalled();
+        $pqb->execute()->willReturn($cursor);
+
+        $this->execute()->shouldReturn($cursor);
+    }
+
+    function it_does_not_add_a_default_filter_on_parents_when_there_is_a_source_id_filter($pqb, CursorInterface $cursor)
+    {
+        $pqb->getRawFilters()->willReturn(
+            [
+                [
+                    'field'    => 'id',
+                    'operator' => 'IN LIST',
+                    'value'    => ['toto'],
+                    'context'  => [],
+                    'type'     => 'field'
+                ],
+            ]
+        );
+
+        $pqb->addFilter('parent', Argument::cetera())->shouldNotBeCalled();
+        $pqb->execute()->willReturn($cursor);
+
+        $this->execute()->shouldReturn($cursor);
+    }
+
+    function it_does_not_add_a_default_filter_on_parents_when_there_is_a_source_identifier_filter($pqb, CursorInterface $cursor)
+    {
+        $pqb->getRawFilters()->willReturn(
+            [
+                [
+                    'field'    => 'identifier',
+                    'operator' => 'IN LIST',
+                    'value'    => ['toto'],
+                    'context'  => [],
+                    'type'     => 'field'
+                ],
+            ]
+        );
+
+        $pqb->addFilter('parent', Argument::cetera())->shouldNotBeCalled();
+        $pqb->execute()->willReturn($cursor);
+
+        $this->execute()->shouldReturn($cursor);
+    }
+
+    function it_does_not_add_a_default_filter_on_parents_when_there_is_a_source_entity_type_filter($pqb, CursorInterface $cursor)
+    {
+        $pqb->getRawFilters()->willReturn(
+            [
+                [
+                    'field'    => 'entity_type',
+                    'operator' => 'EQUALS',
+                    'value'    => 'toto',
+                    'context'  => [],
+                    'type'     => 'field'
+                ],
+            ]
+        );
+
+        $pqb->addFilter('parent', Argument::cetera())->shouldNotBeCalled();
+        $pqb->execute()->willReturn($cursor);
+
+        $this->execute()->shouldReturn($cursor);
+    }
+
+    function it_does_not_add_a_default_filter_on_parents_when_there_is_a_source_ancestor_filter($pqb, CursorInterface $cursor)
+    {
+        $pqb->getRawFilters()->willReturn(
+            [
+                [
+                    'field'    => 'ancestor.id',
+                    'operator' => 'IN LIST',
+                    'value'    => ['toto'],
+                    'context'  => [],
+                    'type'     => 'field'
+                ],
+            ]
+        );
+
+        $pqb->addFilter('parent', Argument::cetera())->shouldNotBeCalled();
+        $pqb->execute()->willReturn($cursor);
+
+        $this->execute()->shouldReturn($cursor);
+    }
+
+    function it_does_not_add_a_default_filter_on_parents_when_there_is_a_source_ancestor_or_self_filter($pqb, CursorInterface $cursor)
+    {
+        $pqb->getRawFilters()->willReturn(
+            [
+                [
+                    'field'    => 'self_and_ancestor.id',
+                    'operator' => 'IN LIST',
+                    'value'    => ['toto'],
+                    'context'  => [],
+                    'type'     => 'field'
+                ],
+            ]
+        );
+
+        $pqb->addFilter('parent', Argument::cetera())->shouldNotBeCalled();
         $pqb->execute()->willReturn($cursor);
 
         $this->execute()->shouldReturn($cursor);
